pipeline/requirements.txt
openai>=1.0.0
pinecone>=8.0.0
requests>=2.32.0
python-dotenv>=1.0.0
tqdm>=4.67.0
